requirements.txt
streamlit==1.39.0
langchain==0.3.7
langchain-openai==0.2.8
langchain-chroma==0.1.4
langchain-community==0.3.7
langchain-text-splitters==0.3.2
langchain-google-genai==2.0.7
chromadb==0.5.18
python-dotenv==1.0.1
